Ditch f32.

+++ src/color/black_body.rs
@@ -1,13 +1,13 @@
-use crate::Float;
+use crate::f64;
 use simple_si_units::base::{Distance, Temperature};
 
 pub(crate) fn planck_radiant_emittance(
-    wavelength: Distance<Float>,
-    temperature: Temperature<Float>,
-) -> Float {
-    const H: Float = 6.62607015e-34;
-    const C: Float = 299792458.0;
-    const K: Float = 1.380649e-23;
+    wavelength: Distance<f64>,
+    temperature: Temperature<f64>,
+) -> f64 {
+    const H: f64 = 6.62607015e-34;
+    const C: f64 = 299792458.0;
+    const K: f64 = 1.380649e-23;
     let lambda = wavelength.as_meters();
     let t = temperature.as_kelvin();
     let a = 2.0 * H * C * C;
@@ -22,9 +22,9 @@ mod tests {
     use super::*;
 
     struct TestExpectance {
-        wavelength: Distance<Float>,
-        temperature: Temperature<Float>,
-        expected: Float,
+        wavelength: Distance<f64>,
+        temperature: Temperature<f64>,
+        expected: f64,
     }
 
     #[test]
